requirements.txt
streamlit>=1.24.0

pymupdf>=1.23.8
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-text-splitters>=0.2.0
langchain-huggingface>=0.0.1
langchain-google-genai>=1.0.0

langchain-chroma>=0.1.0
chromadb>=0.4.22

redis>=5.0.0
hiredis>=2.2.0

sentence-transformers>=2.6.0

python-dotenv

numpy
scikit-learn
pydantic